rename loader file
https://coveord.atlassian.net/browse/KIT-661
